requirements.txt
torch==2.3.1
opencv-python==4.10.0
pillow==10.3.0
pandas==2.2.2
scipy==1.14.0
scikit-image==0.24.0
tensorboardx==2.6.2
